refactor(android): ol index update and margin update logic separation
