+++ android/src/main/java/com/swmansion/enriched/textinput/watchers/EnrichedSpanWatcher.kt
@@ -27,7 +27,7 @@ class EnrichedSpanWatcher(
     end: Int,
   ) {
     updateNextLineLayout(what, text, end)
-    updateUnorderedListSpans(what, text)
+    updateUnorderedListSpans(what, text, end)
     emitEvent(text, what)
   }
 
@@ -38,7 +38,7 @@ class EnrichedSpanWatcher(
     end: Int,
   ) {
     updateNextLineLayout(what, text, end)
-    updateUnorderedListSpans(what, text)
+    updateUnorderedListSpans(what, text, end)
     emitEvent(text, what)
   }
 
@@ -56,9 +56,10 @@ class EnrichedSpanWatcher(
   private fun updateUnorderedListSpans(
     what: Any,
     text: Spannable,
+    end: Int,
   ) {
     if (what is EnrichedInputOrderedListSpan) {
-      view.listStyles?.updateOrderedListIndexes(text)
+      view.listStyles?.updateOrderedListIndexes(text, end)
     }
   }
